# swagger.yaml
openapi: 3.0.0
info:
  title: Dog API
  description: API for managing dogs.
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
  /api/dogs:
    get:
      summary: List all dogs
      # Añade más detalles como parámetros si es necesario.
      responses:
        200:
          description: A list of dogs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dog"
  # Añade otras rutas aquí.

  /api/dog/add:
    post:
      summary: Add a new dog
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Dog"
      responses:
        "201":
          description: Dog created

  /api/dogs/{dogId}:
    get:
      summary: Get a dog by ID
      parameters:
        - name: dogId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A dog object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dog"
        "404":
          description: Dog not found

    put:
      summary: Update a dog by ID
      parameters:
        - name: dogId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dog"
      responses:
        "200":
          description: Dog updated
        "404":
          description: Dog not found

    delete:
      summary: Delete a dog by ID
      parameters:
        - name: dogId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dog deleted
        "404":
          description: Dog not found

components:
  schemas:
    Dog:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The dog's ID.
        name:
          type: string
          description: The dog's name.
        color:
          type: string
        size:
          type: string
        age:
          type: integer
        gender:
          type: string
        breed:
          type: string
        # Define otras propiedades aquí.